Add user avatars.
They are uploaded asynchronously to S3. Once they have been uploaded, the client sends a request to notify the server where the newly uploaded file was put.

The server then starts enqueues a delayed job to fetch the image, process it, and finally reupload it to its final path on the server.